+++ CHANGELOG.md
@@ -72,6 +72,7 @@
     * Change: Prevent PHP warnings due to missing core-related generated files
     * New: Add tests to verify filtered values output nothing when empty
     * New: Add test to check that the appropriate number of framework strings are translatable
+    * New: Add tests for `help_tabs_content()`
     * Change: In bootstrap, add backcompat for PHPUnit pre-v6.0
     * Fix: Update tests to handle deprecation of core's `print_inline_script()`
     * Change: Explicitly define return types for overridden methods